dcc.c: Make fastsend work with psend. perl.c: Undo the "trying to eval" printing, it's too long. util.c: Fix MHz/GHz printing on win32.
git-svn-id: https://xchat.svn.sourceforge.net/svnroot/xchat@501 893a96be-7f27-4fdf-9d1e-6aeec9d3cce1

+++ src/common/util.c
@@ -522,7 +522,7 @@ get_cpu_str (void)
 	mhz = get_mhz ();
 	if (mhz)
 	{
-		int cpuspeed = ( mhz > 1000 ) ? mhz / 1000 : mhz;
+		double cpuspeed = ( mhz > 1000 ) ? mhz / 1000 : mhz;
 		const char *cpuspeedstr = ( mhz > 1000 ) ? "GHz" : "MHz";
 		sprintf (verbuf, "Windows %ld.%ld [i%d86/%.2f%s]",
 					osvi.dwMajorVersion, osvi.dwMinorVersion, si.wProcessorLevel,
